Find out comment style from file content as a fallback
Usually we find out the language based on file extension and use the
comment style used for the language. This PR lets helix figure out the
correct comment style by peeking the first few lines of the file. This
is a fallback, when helix is unable to figure out the language for any
reason.

fixes #5380